ci(actions): when `installed`, tag verifiers in comment instead of assigning
