+++ .github/workflows/issue-notifications.yml
@@ -11,14 +11,14 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
-      - name: "Assign for verification"
+      - name: "Installed notification"
         if: github.event.label.name == '3 - installed'
         uses: actions/github-script@v7
         env:
           ISSUE_VERIFIERS: ${{secrets.ISSUE_VERIFIERS}}
         with:
           script: |
-            const action = require('${{ github.workspace }}/.github/scripts/assignForVerification.js')
+            const action = require('${{ github.workspace }}/.github/scripts/notifyWhenInstalled.js')
             await action({github, context, core})
 
       - name: "Spike complete notification"
